Clean testing/_utils.py
