+++ python/cudf/cudf/tests/test_udf_masked_ops.py
@@ -5,6 +5,10 @@ import operator
 import numpy as np
 import pytest
 from numba import cuda
+from numba.core.typing import signature as nb_signature
+from numba.core.typing.templates import AbstractTemplate
+from numba.cuda.cudadecl import registry as cuda_decl_registry
+from numba.cuda.cudaimpl import lower as cuda_lower
 
 import cudf
 from cudf.core._compat import PANDAS_CURRENT_SUPPORTED_VERSION, PANDAS_VERSION
@@ -16,13 +20,53 @@ from cudf.core.udf._ops import (
     unary_ops,
 )
 from cudf.core.udf.api import Masked
+from cudf.core.udf.strings_lowering import (
+    cast_string_view_to_managed_udf_string,
+)
+from cudf.core.udf.strings_typing import (
+    StringView,
+    managed_udf_string,
+    string_view,
+)
 from cudf.core.udf.utils import precompiled
 from cudf.testing import assert_eq
 from cudf.testing._utils import (
     _decimal_series,
     parametrize_numeric_dtypes_pairwise,
-    sv_to_managed_udf_str,
 )
+
+
+def sv_to_managed_udf_str(sv):
+    """
+    Cast a string_view object to a managed_udf_string object
+
+    This placeholder function never runs in python
+    It exists only for numba to have something to replace
+    with the typing and lowering code below
+
+    This is similar conceptually to needing a translation
+    engine to emit an expression in target language "B" when
+    there is no equivalent in the source language "A" to
+    translate from. This function effectively defines the
+    expression in language "A" and the associated typing
+    and lowering describe the translation process, despite
+    the expression having no meaning in language "A"
+    """
+    pass
+
+
+@cuda_decl_registry.register_global(sv_to_managed_udf_str)
+class StringViewToUDFStringDecl(AbstractTemplate):
+    def generic(self, args, kws):
+        if isinstance(args[0], StringView) and len(args) == 1:
+            return nb_signature(managed_udf_string, string_view)
+
+
+@cuda_lower(sv_to_managed_udf_str, string_view)
+def sv_to_udf_str_testing_lowering(context, builder, sig, args):
+    return cast_string_view_to_managed_udf_string(
+        context, builder, sig.args[0], sig.return_type, args[0]
+    )
 
 
 @pytest.fixture(scope="module")
