+++ python/cudf/cudf_pandas_tests/test_disable_pandas_accelerator.py
@@ -2,30 +2,26 @@
 
 import os
 import subprocess
+import sys
 
 import pytest
-
-from cudf.testing import _utils as utils
 
 
 @pytest.mark.flaky(reruns=3, reruns_delay=30)
 def test_disable_pandas_accelerator_multi_threaded():
     data_directory = os.path.dirname(os.path.abspath(__file__))
-    # Create a copy of the current environment variables
-    env = os.environ.copy()
 
-    with utils.cudf_timeout(20):
-        sp_completed = subprocess.run(
-            [
-                "python",
-                "-m",
-                "cudf.pandas",
-                data_directory + "/data/disable_cudf_pandas_multi_thread.py",
-            ],
-            capture_output=True,
-            text=True,
-            env=env,
-        )
+    sp_completed = subprocess.run(
+        [
+            sys.executable,
+            "-m",
+            "cudf.pandas",
+            data_directory + "/data/disable_cudf_pandas_multi_thread.py",
+        ],
+        capture_output=True,
+        text=True,
+        timeout=20,
+    )
     assert sp_completed.returncode == 0
     output = sp_completed.stdout
